feat(DB): Removed files and leveldb databases
Files database was not used and only kept for backwards compatibility reasons. As we don't need to keep compatibility code was removed.

LevelDB was an alternative to SQLite database with minor improvements in performance but with possibility of crashing. In general don't see much of a point in keeping it.

Removed references to both, and removed their build options.